8258094: AIX build fails after 8257602
